Extract Overview mapper
