+++ WordPress/src/main/java/org/wordpress/android/ui/stats/refresh/StatsModule.kt
@@ -16,7 +16,7 @@ import org.wordpress.android.ui.stats.refresh.lists.sections.granular.UseCaseFac
 import org.wordpress.android.ui.stats.refresh.lists.sections.granular.usecases.ClicksUseCase.ClicksUseCaseFactory
 import org.wordpress.android.ui.stats.refresh.lists.sections.granular.usecases.PostsAndPagesUseCase.PostsAndPagesUseCaseFactory
 import org.wordpress.android.ui.stats.refresh.lists.sections.granular.usecases.ReferrersUseCase.ReferrersUseCaseFactory
-import org.wordpress.android.ui.stats.refresh.lists.sections.granular.usecases.VisitsAndViewsUseCase.VisitsAndViewsUseCaseFactory
+import org.wordpress.android.ui.stats.refresh.lists.sections.granular.usecases.OverviewUseCase.VisitsAndViewsUseCaseFactory
 import org.wordpress.android.ui.stats.refresh.lists.sections.insights.usecases.AllTimeStatsUseCase
 import org.wordpress.android.ui.stats.refresh.lists.sections.insights.usecases.CommentsUseCase
 import org.wordpress.android.ui.stats.refresh.lists.sections.insights.usecases.FollowersUseCase
